+++ Modules/Intent.Modules.Dapr.AspNetCore/Templates/DaprConfiguration/DaprConfigurationTemplatePartial.cs
@@ -27,7 +27,7 @@ namespace Intent.Modules.Dapr.AspNetCore.Templates.DaprConfiguration
         public DaprConfigurationTemplate(IOutputTarget outputTarget, object model = null) : base(TemplateId, outputTarget, model)
         {
             ExecutionContext.EventDispatcher.Subscribe<DaprServiceRegistration>(_services.Add);
-            AddNugetDependency(NuGetPackages.DaprAspNetCore);
+            AddNugetDependency(NugetPackages.DaprAspNetCore(outputTarget));
 
             CSharpFile = new CSharpFile(this.GetNamespace(), this.GetFolderPath())
                 .IntentManagedFully()